% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot.timeseries}
\alias{plot.timeseries}
\title{Plot Quandl time-series}
\usage{
\method{plot}{timeseries}(time_series, column_names, codes = NULL,
  collapse = "2 weeks", func = NULL, ...)
}
\arguments{
\item{time_series}{List of data.frame retrieved from Quandl Dataset.}

\item{column_names}{String or list of strings of titles for each column in this time-series. Map one or more columns from each time-sereis}

\item{codes}{String or list of strings, code a.k.a ticker symbol on Quandle to be plotted. If NULL, then all codes are being included into a plot.}

\item{collapse}{A string giving the distance between breaks like "2 weeks", or "10 years". Default is 2 weeks.}

\item{func}{If FALSE, STOP is a code for a predefined date range is not found, otherwise WARNIGN. Default=FALSE}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
a ggplot object
}
\description{
Plot time-sereis from the Quandl Dataset.
}
\examples{
\dontrun{
# Plot only only High and Low prices for each code (symbol) 
plot.timeseries (time_series=time_series, column_names=c("High", "Low"), codes=c("FB", "AAPL"))

# Plot only only Open and Close prices for Facebook (FB) and Apple (AAPL)
plot.timeseries (time_series=time_series, column_names=c("Open", "Close"), codes=c("FB", "AAPL"))

# Plot only only Open and Close prices for Facebook (FB) and Apple (AAPL) and add square root function
plot.timeseries (time_series=time_series, column_names=c("Open", "Close"), codes=c("FB", "AAPL"), func=y~sqrt(x))
}
}
