% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_piechart}
\alias{plot_piechart}
\title{Plot Quandl pie chart}
\usage{
plot_piechart(time_series, w, codes = NULL, ...)
}
\arguments{
\item{time_series}{List of data.frame retrieved from Quandl Dataset.}

\item{w}{List of numbers, weights of stocks. If less then 1, "Other" code will be created.}

\item{codes}{String or list of strings, code a.k.a ticker symbol on Quandle to be plotted. If NULL, then all codes are being included into a plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
a ggplot object
}
\description{
Plot pie chart from the Quandl Dataset.
}
\examples{
\dontrun{
# Plot only only Open and Close prices for Facebook (FB) and Apple (AAPL)
plot_piechart (time_series=time_series, w=c(0.3, 0.7), codes=c("FB", "AAPL"))

plot.barchart (time_series=time_series, w=c(0.3, 0.3, 0.3), codes=c("FB", "AAPL", "GOOG"))
}
}
