% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_timeseries_all}
\alias{plot_timeseries_all}
\title{Plot all Quandl time-series}
\usage{
plot_timeseries_all(time_series, column_names, codes = NULL,
  by = "codes", ncol = 2, collapse = "2 weeks", func = NULL, ...)
}
\arguments{
\item{time_series}{List of data.frame retrieved from Quandl Dataset.}

\item{column_names}{String or list of strings of titles for each column in this time-series. Map one or more columns from each time-sereis}

\item{codes}{String or list of strings, code a.k.a ticker symbol on Quandle to be plotted. If NULL, then all codes are being included into a plot.}

\item{by}{A string: code or columns. Split into multuple graps.}

\item{ncol}{Integer, number of column}

\item{collapse}{A string giving the distance between breaks like "2 weeks", or "10 years". Default is 2 weeks.}

\item{func}{If FALSE, STOP is a code for a predefined date range is not found, otherwise WARNIGN. Default=FALSE}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{gen_title}{String. Add automatically generated title.}
}
\value{
a ggplot object
}
\description{
Plot all time-sereis from the Quandl Dataset on the same page.
}
\examples{
\dontrun{
# Plot only only High and Low prices for each code (symbol) 
plot.timeseries (time_series=time_series, column_names=c("High", "Low"), codes=c("FB", "AAPL"))

# Plot only only Open and Close prices for Facebook (FB) and Apple (AAPL)
plot.timeseries (time_series=time_series, column_names=c("Open", "Close"), codes=c("FB", "AAPL"))

# Plot only only Open and Close prices for Facebook (FB) and Apple (AAPL) and add square root function
plot.timeseries (time_series=time_series, column_names=c("Open", "Close"), codes=c("FB", "AAPL"), func=y~sqrt(x))
}
}
